+++ src/main/java/org/vaadin/addon/leaflet/shared/LeafletMapClientRpc.java
@@ -5,7 +5,7 @@ import com.vaadin.shared.communication.ClientRpc;
 
 public interface LeafletMapClientRpc extends ClientRpc {
 	
-	void setCenter(Double lat, Double lon, Integer zoom);
+	void setCenter(Double lat, Double lon, Double zoom);
 	void zoomToExtent(Bounds bounds);
     void setMaxBounds(Bounds bounds);
